signed int for windows openmp

+++ src/Open3D/Core/Kernel/NonZeroCPU.cpp
@@ -63,7 +63,7 @@ Tensor NonZeroCPU(const Tensor& src) {
 #ifdef _OPENMP
 #pragma omp parallel for schedule(static)
 #endif
-    for (size_t i = 0; i < num_non_zeros; i++) {
+    for (int64_t i = 0; i < static_cast<int64_t>(num_non_zeros); i++) {
         int64_t non_zero_index = non_zero_indices[i];
         for (int64_t dim = num_dims - 1; dim >= 0; dim--) {
             *static_cast<int64_t*>(result_iter.GetPtr(
